Questions showing up and show more button in place
